Agent_Cloning/pyproject.toml
[project]
name = "agent-cloning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.2.9",
]
